requirements.txt
Django
djangorestframework
torch>=1.7.1
pydub
transformers
openai
python-dotenv
django-cors-headers
joblib
pyaudio
librosa